raspberrypi/requirements.txt
paho-mqtt==1.6.1
bleak==0.13.0
flask==2.0.2
